+++ frontend/common/src/helpers/dateUtils.ts
@@ -1,0 +1,47 @@
+import { IntlShape } from "react-intl";
+import { Maybe, Scalars } from "../api/generated";
+import { Locales } from "./localize";
+
+export function formattedDate(date: Scalars["Date"], locale: Locales) {
+  const formatter = new Intl.DateTimeFormat(locale, {
+    year: "numeric",
+    month: "short",
+  });
+  const formatDate = formatter.format(new Date(date));
+  const formattedMonth = formatDate.substring(0, 4).toUpperCase();
+  const formattedYear = formatDate.substring(4, 10);
+  return `${formattedMonth}  ${formattedYear}`;
+}
+
+export function getDateRange({
+  endDate,
+  startDate,
+  intl,
+  locale,
+}: {
+  endDate: Maybe<Scalars["Date"]>;
+  startDate: Maybe<Scalars["Date"]>;
+  intl: IntlShape;
+  locale: Locales;
+}): React.ReactNode {
+  if (!startDate) return null;
+  const d1 = formattedDate(startDate, locale);
+  if (!endDate)
+    return intl.formatMessage(
+      {
+        defaultMessage: "Since: {d1}",
+        description: "Since",
+      },
+      { d1 },
+    );
+  const d2 = formattedDate(endDate, locale);
+  return endDate
+    ? `${d1} - ${d2}`
+    : intl.formatMessage(
+        {
+          defaultMessage: "Since: {d1}",
+          description: "Since",
+        },
+        { d1 },
+      );
+}
